+++ programacion/ejercicios_programacion/RestaNoNegativa/Programa12.cs
@@ -7,11 +7,11 @@ namespace RestaNoNegativa {
                         "No puede devolver el resultado como un número negativo.");
 
       Console.Write("Introduce el primer número: ");
-      int numero_01 = Convert.ToInt32(Console.ReadLine());
+      double numero_01 = Convert.ToDouble(Console.ReadLine());
       Console.Write("Introduce el segundo número: ");
-      int numero_02 = Convert.ToInt32(Console.ReadLine());
+      double numero_02 = Convert.ToDouble(Console.ReadLine());
 
-      int numero_resta = numero_01 - numero_02;
+      double numero_resta = numero_01 - numero_02;
 
       if (numero_01 >= numero_02) {
         Console.WriteLine(numero_01 + " - " + numero_02 + " = " + numero_resta);
